+++ libs/cardano-ledger-api/cardano-ledger-api.cabal
@@ -56,18 +56,15 @@ library
         bytestring,
         cardano-ledger-allegra ^>=1.5,
         cardano-ledger-alonzo ^>=1.8,
-        cardano-ledger-babbage ^>=1.8,
+        cardano-ledger-babbage ^>=1.8.1,
         cardano-ledger-binary ^>=1.3,
         cardano-ledger-conway >=1.13 && <1.16,
         cardano-ledger-core ^>=1.13,
         cardano-ledger-mary >=1.5 && <1.7,
         cardano-ledger-shelley ^>=1.12,
-        cardano-slotting,
         containers,
         FailT,
         microlens,
-        plutus-ledger-api,
-        text,
         transformers
 
 library testlib
